backend/requirements.txt
#pip install -r requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
python-dotenv